refactor: 네비게이션 바 컴포넌트 공통으로 분리
네비게이션 바 컴포넌트 공통으로 분리하여 반영함

refs: #21
