+++ src/shared/layouts/BottomNavigationBar/index.ts
@@ -1,0 +1,3 @@
+import BottomNavigationBar from './BottomNavigationBar'
+
+export default BottomNavigationBar
