--- src/shared/layouts/NavigationBar/index.ts
@@ -1,3 +1,0 @@
-import NavigationBar from './NavigationBar'
-
-export default NavigationBar
